feat: full proyect structure rename ; remove unused abstract methods paciente ; changed type nrohistoriaclinica, numeromatricula
