+++ src/mdp/ingenieria/clinicagestion/model/decorator/decoratormedico/DecoratorMedicoContratacion.java
@@ -1,0 +1,15 @@
+package mdp.ingenieria.clinicagestion.model.decorator.decoratormedico;
+
+import mdp.ingenieria.clinicagestion.model.decorator.DecoratorMedico;
+import mdp.ingenieria.clinicagestion.model.persona.IMedico;
+
+
+public abstract class DecoratorMedicoContratacion extends DecoratorMedico 
+{
+
+	public DecoratorMedicoContratacion( IMedico medico ) 
+	{	
+		super( medico );
+	}
+	
+}
