+++ src/mdp/ingenieria/clinicagestion/model/persona/IMedico.java
@@ -1,0 +1,10 @@
+package mdp.ingenieria.clinicagestion.model.persona;
+
+public interface IMedico 
+{
+	int getNumeroMatricula();
+	
+	String getEspecialidad();
+	
+	double getHonorario();
+}
