--- src/sistema/gestion/clinica/Interfaces/IMedico.java
@@ -1,6 +1,0 @@
-package sistema.gestion.clinica.Interfaces;
-
-public interface IMedico {
-    String getNumeroMatricula();
-    double getHonorarios();
-}
