--- src/sistema/gestion/clinica/Interfaces/IProtocoloOcuparSala.java
@@ -1,8 +1,0 @@
-package sistema.gestion.clinica.Interfaces;
-
-public interface IProtocoloOcuparSala {
-    void ocuparSala();
-    void ocuparSalaNino(IProtocoloOcuparSala ocuparSala);
-    void ocuparSalaJoven(IProtocoloOcuparSala ocuparSala);
-    void ocuparSalaMayor(IProtocoloOcuparSala ocuparSala);
-}
