--- src/sistema/gestion/clinica/Main.java
@@ -1,9 +1,0 @@
-package sistema.gestion.clinica;
-
-public class Main {
-
-	public static void main(String[] args) {
-		System.out.println("Entry point...");
-	}
-
-}
