--- src/sistema/gestion/clinica/clinica/HabitacionPrivada.java
@@ -1,19 +1,0 @@
-package sistema.gestion.clinica.clinica;
-
-public class HabitacionPrivada extends Habitacion{
-    public HabitacionPrivada(double costoAsignacion, double costoAdicional)
-    {
-        super(costoAsignacion,costoAdicional);
-    }
-    @Override
-    public double getCostoAdicional(int dias)
-    {
-        double costo = this.costoAdicional;
-        if(dias>=2 && dias<=5)
-            costo*=1.3;
-        else
-            if(dias>6)
-                costo*=2;
-        return costo*dias;
-    }
-}
