--- src/sistema/gestion/clinica/model/IMedico.java
@@ -1,10 +1,0 @@
-package sistema.gestion.clinica.model;
-
-public interface IMedico 
-{
-	String getNumeroMatricula();
-	
-	String getEspecialidad();
-	
-	double getHonorario();
-}
